+++ index.html
@@ -232,5 +232,5 @@
   </template>
 
 </body>
-<script src="js/main.js"></script>
+<script src="js/main.js" type="module"></script>
 </html>
